+++ xdef/src/main/java/org/xdef/impl/XCodeProcessor.java
@@ -47,6 +47,7 @@ import org.xdef.XDParseResult;
 import org.xdef.XDParser;
 import org.xdef.XDPool;
 import org.xdef.XDPrice;
+import org.xdef.XDQName;
 import org.xdef.XDRegex;
 import org.xdef.XDRegexResult;
 import org.xdef.XDReport;
@@ -59,6 +60,8 @@ import org.xdef.XDValue;
 import org.xdef.XDValueAbstract;
 import static org.xdef.XDValueID.XD_ANY;
 import static org.xdef.XDValueID.XD_BIGINTEGER;
+import static org.xdef.XDValueID.XD_BNFGRAMMAR;
+import static org.xdef.XDValueID.XD_BNFRULE;
 import static org.xdef.XDValueID.XD_BOOLEAN;
 import static org.xdef.XDValueID.XD_BYTES;
 import static org.xdef.XDValueID.XD_CONTAINER;
@@ -278,6 +281,9 @@ import static org.xdef.impl.code.CodeTable.GET_PARSED_RESULT;
 import static org.xdef.impl.code.CodeTable.GET_PARSED_STRING;
 import static org.xdef.impl.code.CodeTable.GET_PARSED_VALUE;
 import static org.xdef.impl.code.CodeTable.GET_QNAMEURI;
+import static org.xdef.impl.code.CodeTable.GET_QNAME_LOCALNAME;
+import static org.xdef.impl.code.CodeTable.GET_QNAME_NAMESPACE;
+import static org.xdef.impl.code.CodeTable.GET_QNAME_PREFIX;
 import static org.xdef.impl.code.CodeTable.GET_REGEX_GROUP;
 import static org.xdef.impl.code.CodeTable.GET_REGEX_GROUP_END;
 import static org.xdef.impl.code.CodeTable.GET_REGEX_GROUP_NUM;
@@ -529,7 +535,7 @@ import org.xdef.impl.code.DefQName;
 import org.xdef.impl.code.DefSQLService;
 import org.xdef.impl.code.DefString;
 import org.xdef.impl.code.DefTelephone;
-import org.xdef.impl.code.DefURI;
+import org.xdef.impl.code.DefUri;
 import org.xdef.impl.code.DefXPathExpr;
 import org.xdef.impl.code.DefXQueryExpr;
 import org.xdef.impl.code.DefXmlWriter;
@@ -2380,6 +2386,9 @@ public final class XCodeProcessor {
                     _stack[sp] = new DefLong(((XDContainer)_stack[sp]).getXDItem(index).getItemId());
                     continue;
                 }
+                case GET_QNAME_PREFIX: _stack[sp] = new DefString(((XDQName) _stack[sp]).getPrefix()); continue;
+                case GET_QNAME_LOCALNAME: _stack[sp] = new DefString(((XDQName) _stack[sp]).getLocalName()); continue;
+                case GET_QNAME_NAMESPACE: _stack[sp] = new DefString(((XDQName) _stack[sp]).getNamespace()); continue;
                 case GET_RESULTSET_ITEM: {
                     String s = _stack[sp--].stringValue();
                     XDResultSet it = (XDResultSet) _stack[sp];
@@ -3118,7 +3127,7 @@ public final class XCodeProcessor {
                 case NEW_CURRENCY: _stack[sp] = new XDCurrency(_stack[sp].toString()); continue;
                 case NEW_EMAIL: _stack[sp] = new DefEmailAddr(_stack[sp].toString()); continue;
                 case NEW_IPADDR: _stack[sp] = new DefIPAddr(_stack[sp].toString()); continue;
-                case NEW_URI: _stack[sp] = new DefURI(_stack[sp].toString()); continue;
+                case NEW_URI: _stack[sp] = new DefUri(_stack[sp].toString()); continue;
                 case NEW_PARSER: {
                     XDParser p = getParser(item.stringValue());
                     int np = item.getParam();
@@ -4102,6 +4111,8 @@ public final class XCodeProcessor {
                                         case XD_XPATH: pars[j + k] = _stack[i].stringValue(); break;
                                         case XD_IPADDR:
                                         case XD_CURRENCY: pars[j + k] = _stack[i].getObject(); break;
+                                        case XD_BNFGRAMMAR:
+                                        case XD_BNFRULE:
                                         case XD_REGEX:
                                         case XD_REGEXRESULT:
                                         case XD_INPUT: