+++ xdef/src/main/java/org/xdef/impl/compile/CompileBase.java
@@ -102,15 +102,15 @@ import static org.xdef.impl.code.CodeTable.CONTEXT_GETELEMENT_X;
 import static org.xdef.impl.code.CodeTable.CONTEXT_GETLENGTH;
 import static org.xdef.impl.code.CodeTable.CONTEXT_GETTEXT;
 import static org.xdef.impl.code.CodeTable.CONTEXT_ITEM;
-import org.xdef.impl.code.DefContainer;
-import org.xdef.impl.code.DefLong;
-import org.xdef.impl.code.DefString;
-import org.xdef.impl.code.DefXQueryExpr;
 import static org.xdef.impl.code.CodeTable.CONTEXT_ITEMTYPE;
 import static org.xdef.impl.code.CodeTable.CONTEXT_REMOVEITEM;
 import static org.xdef.impl.code.CodeTable.CONTEXT_REPLACEITEM;
 import static org.xdef.impl.code.CodeTable.CONTEXT_SORT;
 import static org.xdef.impl.code.CodeTable.CONTEXT_TO_ELEMENT;
+import org.xdef.impl.code.DefContainer;
+import org.xdef.impl.code.DefLong;
+import org.xdef.impl.code.DefString;
+import org.xdef.impl.code.DefXQueryExpr;
 import static org.xdef.impl.code.CodeTable.CREATE_ELEMENT;
 import static org.xdef.impl.code.CodeTable.CREATE_ELEMENTS;
 import static org.xdef.impl.code.CodeTable.CURRENCYCODE;
@@ -201,6 +201,9 @@ import static org.xdef.impl.code.CodeTable.GET_PARSED_RESULT;
 import static org.xdef.impl.code.CodeTable.GET_PARSED_STRING;
 import static org.xdef.impl.code.CodeTable.GET_PARSED_VALUE;
 import static org.xdef.impl.code.CodeTable.GET_QNAMEURI;
+import static org.xdef.impl.code.CodeTable.GET_QNAME_LOCALNAME;
+import static org.xdef.impl.code.CodeTable.GET_QNAME_NAMESPACE;
+import static org.xdef.impl.code.CodeTable.GET_QNAME_PREFIX;
 import static org.xdef.impl.code.CodeTable.GET_REGEX_GROUP;
 import static org.xdef.impl.code.CodeTable.GET_REGEX_GROUP_END;
 import static org.xdef.impl.code.CodeTable.GET_REGEX_GROUP_NUM;
@@ -428,8 +431,8 @@ public class CompileBase implements CodeTable, XDValueID {
         setType(XD_CURRENCY, "Currency", java.util.Currency.class);
         setType(XD_GPSPOSITION, "GPSPosition", org.xdef.XDGPSPosition.class);
         setType(XD_PRICE, "Price",org.xdef.XDPrice.class);
-        setType(XD_ANYURI, "URI", java.net.URI.class);
-        setType(XD_QNAME, "QName", javax.xml.namespace.QName.class);
+        setType(XD_ANYURI, "XDUri", java.net.URI.class);
+        setType(XD_QNAME, "XDQName", javax.xml.namespace.QName.class);
         setType(XD_EMAIL, "EmailAddr", org.xdef.XDEmailAddr.class);
         setType(XD_IPADDR, "IPAddr", java.net.InetAddress.class);
         setType(XD_TELEPHONE, "Telephone", org.xdef.XDTelephone.class);
@@ -486,7 +489,8 @@ public class CompileBase implements CodeTable, XDValueID {
             ((char) XD_EMAIL) + ";EmailAddr;" +
             ((char) XD_IPADDR) + ";IPAddr;" +
             ((char) XD_TELEPHONE) + ";Telephone;" +
-            ((char) XD_BNFGRAMMAR) + ";DefBNFGrammar;" +
+            ((char) XD_BNFGRAMMAR) + ";BNFGrammar;" +
+            ((char) XD_BNFRULE) + ";BNFRule;" +
             ((char) XD_LOCALE) + ";Locale;" +
             ((char) XD_UNIQUESET_KEY) + ";uniqueSetKey;" +
             ((char) XD_ANY) + ";XDValue;" +
@@ -639,7 +643,7 @@ public class CompileBase implements CodeTable, XDValueID {
             keyParam("pattern",XD_STRING,true,-1,false),
             keyParam("whiteSpace", XD_STRING, false, -1, true, new DefString("collapse"))};
         im = genParserMetnod(0, 2, new short[] {XD_LONG,XD_LONG}, XD_ANYURI, keyPars);
-        parser(im, org.xdef.impl.parsers.XSParseAnyURI.class, "anyURI", "?xs:anyURI");
+        parser(im, org.xdef.impl.parsers.XSParseAnyURI.class, "anyURI", "?xs:anyURI", "?uri" );
         parser(im, org.xdef.impl.parsers.XSParseName.class, "Name", "?xs:Name");
         im = genParserMetnod(0, 2, new short[] {XD_LONG, XD_LONG}, XD_STRING, keyPars);
         parser(im, org.xdef.impl.parsers.XSParseID.class, "ID", "?xs:ID");
@@ -740,16 +744,16 @@ public class CompileBase implements CodeTable, XDValueID {
         parser(im, org.xdef.impl.parsers.XDParseQNameList.class, "QNameList", "?QnameList");
         parser(im, org.xdef.impl.parsers.XDParseLanguages.class, "languages", "?ISOlanguages");
         parser(im, org.xdef.impl.parsers.XDParseCountries.class, "countries");
-        im = genParserMetnod(0, 1, new short[] {XD_STRING}, XD_CONTAINER,
-            keyParam("argument", XD_ANY, true, 1,false),
-                keyParam("enumeration", XD_STRING, true, -1,false),
-                keyParam("length", XD_LONG, false, -1,false),
-                keyParam("maxLength", XD_LONG, false, -1,false),
-                keyParam("minLength", XD_LONG, false, -1,false),
-                keyParam("pattern",XD_STRING,true,-1,false),
-                keyParam("separator", XD_STRING, true,  0, false),
-                keyParam("whiteSpace", XD_STRING, false, -1, true, new DefString("collapse")));
-        parser(im, org.xdef.impl.parsers.XDParseQNameURIList.class,"QNameURIList", "?QnameListURI");
+//        im = genParserMetnod(0, 1, new short[] {XD_STRING}, XD_CONTAINER,
+//            keyParam("argument", XD_ANY, true, 1,false),
+//                keyParam("enumeration", XD_STRING, true, -1,false),
+//                keyParam("length", XD_LONG, false, -1,false),
+//                keyParam("maxLength", XD_LONG, false, -1,false),
+//                keyParam("minLength", XD_LONG, false, -1,false),
+//                keyParam("pattern",XD_STRING,true,-1,false),
+//                keyParam("separator", XD_STRING, true,  0, false),
+//                keyParam("whiteSpace", XD_STRING, false, -1, true, new DefString("collapse")));
+//        parser(im, org.xdef.impl.parsers.XDParseQNameURIList.class,"QNameURIList", "?QnameListURI");
         im = genParserMetnod(0, 2, new short[] {XD_DURATION,XD_DURATION}, XD_DURATION,
             keyParam("base", XD_STRING, true, -1,false),
             keyParam("enumeration", XD_DURATION, true,-1,false),
@@ -850,8 +854,8 @@ public class CompileBase implements CodeTable, XDValueID {
         parser(im,org.xdef.impl.parsers.XDParseFile.class, "file");
         parser(im,org.xdef.impl.parsers.XDParseFileList.class,"fileList");
         parser(im,org.xdef.impl.parsers.XDParseXDType.class, "xdType");
-        parser(im,org.xdef.impl.parsers.XDParseUri.class, "uri");
-        parser(im,org.xdef.impl.parsers.XDParseUriList.class, "uriList");
+//        parser(im,org.xdef.impl.parsers.XDParseUri.class, "uri");
+//        parser(im,org.xdef.impl.parsers.XDParseUriList.class, "uriList");
         parser(im,org.xdef.impl.parsers.XDParseUrl.class, "url");
         parser(im,org.xdef.impl.parsers.XDParseUrlList.class, "urlList");
         parser(im,org.xdef.impl.parsers.XDParseDomainAddr.class, "domainAddr");
@@ -967,12 +971,12 @@ public class CompileBase implements CodeTable, XDValueID {
         method(ti, genInternalMethod(NEW_BNFGRAMAR, XD_BNFGRAMMAR, ANY_MODE,1,2,XD_STRING,XD_BNFGRAMMAR),"#");
         method(ti, genInternalMethod(BNF_PARSE, XD_PARSERESULT,
             ANY_MODE, 2, 3, XD_BNFGRAMMAR, XD_STRING, XD_STRING), "parse", "?check");
-        method(ti, genInternalMethod(GET_BNFRULE, XD_BNFRULE, ANY_MODE, 2,2, XD_BNFGRAMMAR,XD_STRING),"rule");
+        method(ti, genInternalMethod(GET_BNFRULE, XD_BNFRULE, ANY_MODE, 2,2, XD_BNFGRAMMAR,XD_STRING), "rule");
 ////////////////////////////////////////////////////////////////////////////////
 // BNF RULE
 ////////////////////////////////////////////////////////////////////////////////
         ti = XD_BNFRULE;
-        method(ti, genInternalMethod(BNFRULE_PARSE, XD_PARSERESULT,ANY_MODE, 1, 2,XD_BNFRULE,XD_ANY),"parse","?check");
+        method(ti, genInternalMethod(BNFRULE_PARSE, XD_PARSERESULT,ANY_MODE, 1, 2,XD_BNFRULE,XD_ANY), "parse","?check");
         method(ti, genInternalMethod(BNFRULE_VALIDATE, XD_BOOLEAN, ANY_MODE, 1, 2, XD_BNFRULE, XD_ANY), "validate");
 ////////////////////////////////////////////////////////////////////////////////
 // BYTES (array)
@@ -1173,8 +1177,7 @@ public class CompileBase implements CodeTable, XDValueID {
         method(ti, genInternalMethod(GET_REPORT, XD_REPORT, ANY_MODE, 1, 1, XD_OUTPUT), "getLastError");
         method(ti, genInternalMethod(OUT1_STREAM, XD_VOID, ANY_MODE, 2, 2, XD_OUTPUT, XD_STRING), "out");
         method(ti, genInternalMethod(OUTLN1_STREAM, XD_VOID, ANY_MODE, 1, 2, XD_OUTPUT, XD_STRING), "outln");
-        method(ti, genInternalMethod(PRINTF_STREAM, XD_VOID,
-            ANY_MODE, 2, Integer.MAX_VALUE, XD_OUTPUT, XD_ANY), "printf");
+        method(ti, genInternalMethod(PRINTF_STREAM, XD_VOID, ANY_MODE, 2, Integer.MAX_VALUE,XD_OUTPUT,XD_ANY),"printf");
         method(ti, genInternalMethod(PUT_REPORT, XD_VOID, ANY_MODE, 2, 2, XD_OUTPUT, XD_REPORT), "putReport");
 ////////////////////////////////////////////////////////////////////////////////
 // PARSER
@@ -1186,6 +1189,9 @@ public class CompileBase implements CodeTable, XDValueID {
 ////////////////////////////////////////////////////////////////////////////////
         ti = XD_QNAME;
         method(ti, genInternalMethod(NEW_QNAME, XD_QNAME, ANY_MODE, 1, 3, XD_STRING,XD_STRING,XD_STRING), "#");
+        method(ti, genInternalMethod(GET_QNAME_PREFIX, XD_STRING, ANY_MODE, 1, 1, XD_QNAME), "getPrefix");
+        method(ti, genInternalMethod(GET_QNAME_LOCALNAME, XD_STRING, ANY_MODE, 1, 1, XD_QNAME), "getLocalName");
+        method(ti, genInternalMethod(GET_QNAME_NAMESPACE, XD_STRING, ANY_MODE, 1, 1, XD_STRING), "getNamespace");
 ////////////////////////////////////////////////////////////////////////////////
 // PARSERESULT (result of parsing by parsers)
 ////////////////////////////////////////////////////////////////////////////////